requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
Authlib==1.3.1
bandit==1.7.9
black==22.12.0
boto3==1.34.86
botocore==1.34.86
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cryptography==43.0.0
distlib==0.3.8
dparse==0.6.4b0
filelock==3.12.4
flake8==7.1.1
freezegun==1.4.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
moto==5.0.5
mypy-extensions==1.0.0
numpy==2.0.1
packaging==24.0
pathspec==0.12.1
pbr==6.0.0
pg8000==1.31.2
platformdirs==4.2.2
pluggy==1.4.0
psutil==6.0.0
py-partiql-parser==0.5.4
pyarrow==17.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.1.1
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
responses==0.25.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safety==3.2.5
safety-schemas==0.0.3
scramp==1.4.5
shellingham==1.5.4
six==1.16.0
stevedore==5.2.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
Werkzeug==3.0.3
xmltodict==0.13.0
hypothesis
meson
memory_profiler
pytest-cov
pandas
